+++ server/typedefs/account.js
@@ -3,19 +3,19 @@ const { gql } = require('apollo-server');
 module.exports = {
     typeDef: gql`
     type Account {
-        balance:Int!
+        balance:Float!
         owner:User!
         id:ID!
     }
 
     extend type Query{
-        balance:Int!
+        balance:Float!
     }
     
     extend type Mutation{
         createAccount(balance:Int!):Account
-        deposit(amount:Int!):Account
-        spend(amount:Int!):Account        
+        deposit(amount:Float!):Account
+        spend(amount:Float!):Account        
     }
     extend  type Subscription{
         balanceChanged:Account!
